+++ src/main/java/com/almostreliable/summoningrituals/data/SummoningLang.java
@@ -43,7 +43,6 @@ public final class SummoningLang extends LanguageProvider {
     // conditions
     public static final LangEntry ALTAR_PROPERTIES = LangEntry.condition("altar_properties", "Altar Properties");
     public static final LangEntry BIOMES = LangEntry.condition("biomes", "Biomes");
-    public static final LangEntry BLOCK_BELOW = LangEntry.condition("block_below", "Block below");
     public static final LangEntry CLEAR = LangEntry.condition("weather_clear", "Clear");
     public static final LangEntry DIMENSION = LangEntry.condition("dimension", "Dimension");
     public static final LangEntry HEIGHT = LangEntry.condition("height", "Height");