+++ src/main/java/com/almostreliable/summoningrituals/recipe/condition/LocationCondition.java
@@ -3,8 +3,6 @@ package com.almostreliable.summoningrituals.recipe.condition;
 import com.almostreliable.summoningrituals.data.SummoningLang;
 import com.almostreliable.summoningrituals.util.RawHolderSet;
 
-import net.minecraft.advancements.critereon.BlockPredicate;
-import net.minecraft.advancements.critereon.LocationPredicate;
 import net.minecraft.advancements.critereon.MinMaxBounds;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.HolderSet;
@@ -41,16 +39,8 @@ public class LocationCondition implements ConditionHandler<LocationCheck> {
     public void getTooltip(List<Component> tooltip, LocationCheck condition) {
         var opt = condition.predicate();
         if (opt.isEmpty()) return;
+
         var predicate = opt.get();
-
-        if (condition.offset().equals(BlockPos.ZERO)) {
-            getTooltipForAltar(tooltip, predicate);
-        } else if (condition.offset().equals(BlockPos.ZERO.below()) && predicate.block().isPresent()) {
-            getTooltipForBlockBelow(tooltip, predicate.block().get());
-        }
-    }
-
-    private void getTooltipForAltar(List<Component> tooltip, LocationPredicate predicate) {
         if (predicate.biomes().isPresent()) {
             tooltip.add(conditionNameComponent(SummoningLang.BIOMES.get()));
             addHolderSetTooltip(tooltip, predicate.biomes().get());
@@ -82,19 +72,6 @@ public class LocationCondition implements ConditionHandler<LocationCheck> {
             tooltip.add(conditionNameComponent(SummoningLang.STRUCTURES.get()));
             addHolderSetTooltip(tooltip, predicate.structures().get());
         }
-    }
-
-    private void getTooltipForBlockBelow(List<Component> tooltip, BlockPredicate predicate) {
-        var blocksOpt = predicate.blocks();
-        if (blocksOpt.isEmpty()) return;
-
-        var blockName = blocksOpt.get().get(0).value().getName();
-        tooltip.add(conditionNameValueComponent(SummoningLang.BLOCK_BELOW.get(), blockName));
-
-        var propertiesOpt = predicate.properties();
-        if (propertiesOpt.isEmpty()) return;
-        var properties = propertiesOpt.get().properties();
-        addBlockStateTooltip(tooltip, properties);
     }
 
     private void addHeightTooltip(List<Component> tooltip, MinMaxBounds.Doubles yPos) {